moved HaltedTraverserStrategy to the general strategy package as it can now be used with OLTP.
